DELETE /tag now deletes NoteTag connections

+++ scripts/tag.lua
@@ -145,6 +145,12 @@ local function handleDELETE(db)
     if not rows() then return ngx.HTTP_UNAUTHORIZED end
 
     local sql = [[
+        DELETE FROM NoteTag
+            WHERE TagID = ?]]
+    local err = us.sexec(db, sql, tid)
+    if err ~= sqlite3.OK then return ngx.HTTP_INTERNAL_SERVER_ERROR end
+
+    local sql = [[
         DELETE FROM Tags
             WHERE TagID = ? AND UserID = ?]]
     local err = us.sexec(db, sql, tid, uid)
